requirements.txt
cycler==0.10.0
ftfy==6.0.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
scikit-learn
six==1.16.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.2
wcwidth==0.2.5
